refactor(rolldown_plugin_manifest): move utils logic to utils.rs
